ICU-3222 compiler warning fix
X-SVN-Rev: 14313

+++ icu4c/source/i18n/unicode/ucoleitr.h
@@ -25,7 +25,7 @@
  * to be returned.
  * @stable ICU 2.0
  */
-#define UCOL_NULLORDER        0xFFFFFFFF
+#define UCOL_NULLORDER        ((int32_t)0xFFFFFFFF)
 
 #include "unicode/ucol.h"
